refactor: extract toble of contents logic into useToc custom hook
